added UserProfile and Doctor profile
