+++ src/Components/Layout.jsx
@@ -3,6 +3,7 @@ import NavbarComponent from './NavbarComponent'
 // import Footer from './Footer'
 import { Outlet } from 'react-router-dom'
 import { MyContext } from './ContextProvider'
+import Loading from './Loading';
 // import { useAuthState } from 'react-firebase-hooks/auth';
 
 // import { onAuthStateChanged } from 'firebase/auth';
@@ -53,21 +54,7 @@ export default function Layout() {
   return (
     <>
      {pending ? (
-         <div className='loading'>
-         <div id="loading">
-           <div className="sk-cube-grid">
-             <div className="sk-cube sk-cube1" />
-             <div className="sk-cube sk-cube2" />
-             <div className="sk-cube sk-cube3" />
-             <div className="sk-cube sk-cube4" />
-             <div className="sk-cube sk-cube5" />
-             <div className="sk-cube sk-cube6" />
-             <div className="sk-cube sk-cube7" />
-             <div className="sk-cube sk-cube8" />
-             <div className="sk-cube sk-cube9" />
-           </div>
-         </div>
-       </div>
+      <Loading />
       ) : (
         <>
       <div className='tall GreenishBG'>
